requirements.txt
matplotlib==3.3.2
numpy==1.19.4
streamlit==0.71.0
pandas==1.1.4
Pillow==8.0.1
scikit-learn==0.22.1
